*: address staticcheck lints
Addresses the following lint checks:

- S1008: Simplify returning boolean expression.
- S1016: Use a type conversion instead of manually copying struct
  fields.
- S1031: Omit redundant nil check around loop.
- SA4006: A value assigned to a variable is never read before being
  overwritten.
- SA5011: Possible nil pointer dereference.
- SA9003: Empty body in an if or else branch.
- SA9004: Only the first constant has an explicit type.

Addresses #1327.

+++ internal/batchskl/skl_test.go
@@ -111,7 +111,7 @@ type testStorage struct {
 
 func (d *testStorage) add(key string) uint32 {
 	offset := uint32(len(d.data))
-	d.data = append(d.data, base.InternalKeyKindSet)
+	d.data = append(d.data, uint8(base.InternalKeyKindSet))
 	var buf [binary.MaxVarintLen64]byte
 	n := binary.PutUvarint(buf[:], uint64(len(key)))
 	d.data = append(d.data, buf[:n]...)
@@ -121,7 +121,7 @@ func (d *testStorage) add(key string) uint32 {
 
 func (d *testStorage) addBytes(key []byte) uint32 {
 	offset := uint32(len(d.data))
-	d.data = append(d.data, base.InternalKeyKindSet)
+	d.data = append(d.data, uint8(base.InternalKeyKindSet))
 	var buf [binary.MaxVarintLen64]byte
 	n := binary.PutUvarint(buf[:], uint64(len(key)))
 	d.data = append(d.data, buf[:n]...)
